[fix] #43: 상세페이지 댓글 수정, 삭제 시 발생하는 오류 수정

+++ src/main/webapp/board/deleteReply.jsp
@@ -1,6 +1,7 @@
-<%@ page import="java.util.*, java.sql.*, com.bfc.board.ReplyDTO, com.bfc.board.ReplyDAO" %>
+<%@ page import="java.util.*, java.sql.*" %>
 <%@ page language="java" contentType="text/html; charset=UTF-8"
     pageEncoding="UTF-8"%>
+<jsp:useBean id="replyDAO" class="com.bfc.board.ReplyDAO" />
 <!DOCTYPE html>
 <html>
 <head>
@@ -9,12 +10,16 @@
 </head>
 <body>
 	<%
+		String postId = request.getParameter("postId");
 		String replyIdStr = request.getParameter("replyId");
-	    if (replyIdStr != null && !replyIdStr.trim().isEmpty()) {
+		int writerId = Integer.parseInt(request.getParameter("writerId"));
+        int userId = (Integer) (session.getAttribute("userId") != null? session.getAttribute("userId"): -1);
+        boolean isWriter = writerId == userId;
+		
+	    if (isWriter && replyIdStr != null && !replyIdStr.trim().isEmpty()) {
 	        int replyId = Integer.parseInt(replyIdStr);
-	        ReplyDAO replyDAO = new ReplyDAO();
 	        replyDAO.deleteReply(replyId);
-	        response.sendRedirect("boardDetail.jsp");
+	        response.sendRedirect("boardDetail.jsp?postId=" + postId);
 	    } else {
 	        out.println("<script>alert('잘못된 요청입니다.'); history.back();</script>");
 	    }
